% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaFromPointsOrPoly.r
\name{areaFromPointsOrPoly}
\alias{areaFromPointsOrPoly}
\title{Area of a spatial polygon or set of points}
\usage{
areaFromPointsOrPoly(x)
}
\arguments{
\item{x}{Any of: SpatialPoints, SpatialPointsDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame. Must be in an equal-area projection!}
}
\value{
Numeric (area in km2).
}
\description{
This function returns the area of a SpatialPolygon or SpatialPolygonsDataFrame object \emph{or} of the minimum convex polygon of a set of points. Input can be a set of coordinates or a polygon.
}
