% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareResponse.r
\name{compareResponse}
\alias{compareResponse}
\title{Compare two response curves along one or more predictors}
\usage{
compareResponse(pred1, pred2, data, predictor = names(data),
  adjust = FALSE, gap = Inf, smooth = FALSE, smoothN = 1000,
  smoothRange = c(0, 1), graph = FALSE, ...)
}
\arguments{
\item{pred1}{Numeric list. Predictions from first model along \code{data} (one value per row in \code{data}).}

\item{pred2}{Numeric list. Predictions from second model along \code{data} (one value per row in \code{data}).}

\item{data}{Data frame or matrix corresponding to \code{pred1} and \code{pred2}.}

\item{predictor}{Character list. Name(s) of predictor(s) for which to calculate comparisons. These must appear as column names in \code{data}.}

\item{adjust}{Logical. If TRUE then subtract the mean of \code{pred1} from \code{pred1} and the mean of \code{pred2} from \code{pred2} before analysis. Useful for comparing the shapes of curves while controlling for different elevations (intercepts).}

\item{gap}{Numeric >0. Proportion of range of predictor variable across which to assume a gap exists. Calculation of \code{areaAbsDiff} will  ignore gaps wide than this. To ensure the entire range of the data is included set this equal to \code{Inf} (default).}

\item{smooth}{Logical. If TRUE then the responses are first smoothed using loess() then compared at \code{smoothN} values along each predictor. If FALSE then comparisons are conducted at the raw values \code{pred1} and \code{pred2}.}

\item{smoothN}{\code{NULL} or positive integer. Number of values along "pred" at which to calculate comparisons. Only used if \code{smooth} is \code{TRUE}. If \code{NULL}, then comparisons are calculated at each value in data. If a number, then comparisons are calcualted at \code{smoothN} values of \code{data[ , pred]} that cover the range of \code{data[ , pred]}.}

\item{smoothRange}{2-element numeric list or \code{NULL}. If \code{smooth} is TRUE, then force loess'ed predictions < \code{smoothRange[1]} to equal \code{smoothRange[1]} and predictions > \code{smoothRange[2]} to equal \code{smoothRange[2]}. Ignored if \code{NULL}.}

\item{graph}{Logical. If \code{TRUE} then plot predictions.}

\item{...}{Arguments to pass to functions like \code{sum()} (for example, \code{na.rm=TRUE}) and to \code{overlap()} (for example, \code{w} for weights). Note that if \code{smooth} = TRUE then passing an argument called \code{w} will likely cause a warning and make results circumspect \emph{unless} weights are pre-calculated for each of the \code{smoothN} points along a particular predictor.}
}
\value{
Either a data frame (if \code{smooth = FALSE} or a list object with the smooth model plus a data frame (if \code{smooth = TRUE}) . The data frame represents metrics comparing response curves of \code{pred1} and \code{pred2}:
\itemize{
\item \code{predictor} Predictor for which comparison was made
\item \code{n} Number of values of predictor at which comparison was calculated
\item \code{adjust} \code{adjust} argument.
\item \code{smooth} \code{smooth} argument.
\item \code{meanDiff} Mean difference between predictions of \code{pred1} and \code{pred2} (higher ==> more different).
\item \code{meanAbsDiff} Mean absolute value of difference between predictions of \code{pred1} and \code{pred2} (higher ==> more different).
\item \code{areaAbsDiff} Sum of the area between curves predicted by \code{pred1} and \code{pred2}, standardized by total potential area between the two curves to be in the range [0, 1] (higher ==> more different).
\item \code{d} Schoener's \emph{D} (Warren, D.L., Glor, R.E., and Turelli, M.  2008.  Environmental niche equivalency versus conservatism: Quantitative approaches to niche evolution.  \emph{Evolution} 62:2868-2883.)
\item \code{i} Hellinger's \emph{I} (adjusted to have a range [0, 1]; Warren, D.L., Glor, R.E., and Turelli, M.  2008.  Environmental niche equivalency versus conservatism: Quantitative approaches to niche evolution.  |emph{Evolution} 62:2868-2883.)
\item \code{esp} Godsoe's ESP (Godsoe, W. and Case, B.S.  2015.  Accounting for shifts in the frequency of suitable environments when testing for niche overlap.  \emph{Methods in Ecology and Evolution} 6:59-66.)
\item \code{rho} Pearson correlation between predictions of \code{pred1} and \code{pred2}.
\item \code{rankCor} Spearman rank correlation between predictions of \code{pred1} and \code{pred2}.
}
}
\description{
This function calculates a suite of metrics reflecting of niche overlap for two response curves. Response curves are predicted responses of a uni- or multivariate model along a single variable. Depending on the user-specified settings the function calculates these values either at each pair of values of \code{pred1} and \code{pred2} \emph{or} along a smoothed version of \code{pred1} and \code{pred2}.
}
\examples{
set.seed(123)
pred1 <- seq(0.4, 1, length.out=100) * runif(100)
pred2 <- seq(0, 1, length.out=100) * runif(100)
 data <- data.frame(x1=1:100, x2=1:100 + (1:100 - 51)^2)
 compareResponse(pred1, pred2, data)
 compareResponse(pred1, pred2, data, smooth=TRUE)
 compareResponse(pred1, pred2, data, adjust=TRUE)
}
\seealso{
[compareNiches()]
}
