% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contBoyce.r
\name{contBoyce}
\alias{contBoyce}
\title{The Continuous Boyce Index (CBI) with optional weighting.}
\usage{
contBoyce(pres, bg, numBins = 101, binWidth = 0.1, presWeight = rep(1,
  length(pres)), bgWeight = rep(1, length(bg)), autoWindow = TRUE,
  method = "spearman", dropZeros = TRUE, graph = FALSE, na.rm = FALSE)
}
\arguments{
\item{pres}{Numeric list. Predicted values at test presences}

\item{bg}{Numeric list.  Predicted values at background sites.}

\item{numBins}{Positive integer. Number of (overlapping) bins into which to divide predictions.}

\item{binWidth}{Positive numeric. Size of a bin. Each bin will be \code{binWidth * (max - min)} where \code{max} and \code{min}. If \code{autoWindow} is \code{FALSE} (the default) then \code{min} is 0 and \code{max} is 1. If \code{autoWindow} is \code{TRUE} then \code{min} and \code{max} are the maximum and minimum value of all predictions in the background and presence sets (i.e., not necessarily 0 and 1).}

\item{presWeight}{Numeric vector same length as \code{pres}. Relative weights of presence sites.}

\item{bgWeight}{Numeric vector same length as \code{bg}. Relative weights of background sites.}

\item{autoWindow}{Logical. If \code{FALSE} calculate bin boundaries starting at 0 and ending at 1 (default). If \code{TRUE} then calculate bin boundaries starting at minimum predicted value and ending at maximum predicted value.}

\item{method}{Character. Type of correlation to calculate. The default is \code{'spearman'}, the Spearman rank correlation coefficient used by Boyce et al. (2002) and Hirzel et al. (2006), which is the traditional CBI. In contrast, \code{'pearson'} or \code{'kendall'} can be used instead.  See [stats::cor()] for more details.}

\item{dropZeros}{Logical. If TRUE then drop all bins in which the frequency of presences is 0.}

\item{graph}{Logical. If TRUE then plot P vs E and P/E versus bin.}

\item{na.rm}{Logical. If \code{TRUE} then remove any presences and associated weights and background predictions and associated weights with \code{NA}s.}
}
\value{
Numeric value.
}
\description{
This function calculates the Continuous Boyce Index (CBI), a measure of model accuracy for presence-only test data.
}
\details{
The CBI is the Spearman rank correlation coefficient between the proportion of sites in each prediction class and the expected proportion of predictions in each prediction class based on the proportion of the landscape that is in that class. Values >0 indicate the model's output is positively correlated with the true probability of presence.  Values <0 indicate it is negatively correlated with the true probabilty of presence.
}
\examples{
set.seed(123)
pres <- sqrt(runif(100))
bg <- runif(1000)
contBoyce(pres, bg)
contBoyce2x(pres, bg)
presWeight <- c(rep(1, 10), rep(0.5, 40))
contBoyce(pres, bg, presWeight=presWeight)
contBoyce2x(pres, bg, presWeight=presWeight)
\dontrun{
# compare stability of CBI calculated with ecospat.boyce() in ecospat package
library(ecospat)
set.seed(123)
results <- data.frame()
for (perform in c(1, 1.5, 2)) {
	for (i in 1:30) {

		pres <- runif(100)^(1 / perform)
		bg <- runif(1000)

		cbi_enmSdm <- contBoyce(pres, bg)
		cbi_ecospat <- ecospat.boyce(bg, pres, PEplot=FALSE)$Spearman.cor

		results <- rbind(
			results,
			data.frame(
				performance = rep(perform, 2),
				method = c('enmSdm', 'ecospat'),
				cbi = c(cbi_enmSdm, cbi_ecospat)
			)
		)

	}

}

results$performance[results$performance == 1] <- 'poor'
results$performance[results$performance == 2] <- 'OK'
results$performance[results$performance == 3] <- 'good'

results$category <- paste0(results$method, '\\n', results$performance)

par(mfrow=c(1, 2))
boxplot(cbi ~ category,
	data=results,
	ylab='CBI',
	main='CBI of poor, OK, and good models',
	border=c(rep('darkred', 3),
	rep('darkblue', 3))
)
plot(results$cbi,
	pch=rep(c(21, 22, 23, 24), each=2),
	bg=ifelse(results$method == 'ecospat', 'darkred', 'cornflowerblue'),
	main='Pairs of CBIs',
	ylab='CBI'
)
legend('bottomright', fill=c('darkred', 'cornflowerblue'), legend=c('ecospat', 'enmSdm'))
}
}
\references{
Boyce, M.S., Vernier, P.R., Nielsen, S.E., and Schmiegelow, F.K.A.  2002.  Evaluating resource selection functions.  \emph{Ecological Modeling} 157:281-300)

Hirzel, A.H., Le Lay, G., Helfer, V., Randon, C., and Guisan, A.  2006.  Evaluating the ability of habitat suitability models to predict species presences.  \emph{Ecological Modeling} 199:142-152.
}
\seealso{
\code{\link[stats]{cor}}, \code{\link{Fpb}}, \code{\link{aucWeighted}}, \code{link[enmSdm]{contBoyce2x}}
}
