% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTropicosCoords.r
\name{convertTropicosCoords}
\alias{convertTropicosCoords}
\title{Convert coordinates from TROPICOS records to decimal degrees}
\usage{
convertTropicosCoords(long, lat, returnApprox = FALSE)
}
\arguments{
\item{long}{Character list, longitudes in TROPICOS format.}

\item{lat}{Character list, latitudes in TROPICOS format.}

\item{returnApprox}{Logical, if \code{TRUE} then approximate coordinates (surrounded by square brackets in TROPICOS format) are converted. If \code{FALSE} (default), then \code{NA} is returned.}
}
\value{
A data five-column frame. The first two columns are the original TROPICOS coordinates, the next two the converted coordinates, and the last a flag to indicate if the TROPICOS coordinates were approximate (\code{TRUE}) or "exact" (\code{FALSE}).
}
\description{
The TROPICOS (http://www.tropicos.org/) plant specimen database of the Missouri Botanical Garden returns records in degrees-minutes-seconds format with symbols for degrees, minutes, second, and hemisphere.  This function converts this format to signed decimal degrees format.  By default, values surrounded by square brackets (\code{[]}) are returned as \code{NA} because they are geolocated to the center of a political unit (usually county or equivalent).
}
\examples{
long <- c("078°14'05\\"\\"W", "091°49'39\\"\\"W", NA, "[091°53'38\\"\\"W]")
lat <- c("39°40'41\\"\\"N", "36°34'14\\"\\"S", NA, "[36°46'14\\"\\"N]")
convertTropicosCoords(long, lat)
convertTropicosCoords(long, lat, TRUE)
}
\seealso{
\code{\link[enmSdm]{dmsToDecimal}}
}
