% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryISO3.r
\docType{data}
\name{countryISO3}
\alias{countryISO3}
\title{Crosswalk between country names and ISO3 codes}
\format{
An object of class \code{'data.frame'}.
}
\source{

}
\usage{
data(countryISO3)
}
\description{
Data frame of country names and ISO3 codes
}
\examples{
data(countryISO3)
countryISO3
}
\references{

}
\keyword{ISO3,}
\keyword{countries}
\keyword{country,}
\keyword{datasets,}
