% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlotPoly.r
\name{createPlotPoly}
\alias{createPlotPoly}
\title{Create SpatialPolygon same size as a plot}
\usage{
createPlotPoly(x = NULL)
}
\arguments{
\item{Either}{\code{NULL} or an object of class \code{CRS}, a coordinate reference string (proj4 string), or an object with a CRS. If any of these is provided, the \code{SpatialPolygon} object will have this CRS.}
}
\value{
\code{SpatialPolygon}
}
\description{
This function creates a "rectangular" \code{SpatialPolygon} object with the same dimensions as a plot window. It is especially useful for cropping subsequent rasters or \code{SpatialPolygon} objects that would benefit from cropping. A plot must be made before calling this function.
}
\examples{
data(mad0)
plot(mad0)
poly <- createPlotPoly(mad0)
plot(poly, border='blue')
plot(mad0, add=TRUE)

}
