% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddmmssToDecimal.r
\name{ddmmssToDecimal}
\alias{ddmmssToDecimal}
\title{Convert geographic coordinates in degrees-minutes-second to decimal format}
\usage{
ddmmssToDecimal(dd, mm, ss, eastNorth)
}
\arguments{
\item{dd}{Numeric. Degrees longitude or latitude. Can be a decimal value.}

\item{mm}{Numeric. Minutes longitude or latitude. Can be a decimal value.}

\item{ss}{Numeric. Second longitude or latitude. Can be a decimal value.}

\item{eastNorth}{Logical. Set to TRUE if the location is east of the Prime Meridian (longitude) or north of the equator (latitude).}
}
\value{
Numeric.
}
\description{
This function converts geographic coordinates in degrees-minutes-seconds (DD-MM-SS) format to decimal format.
}
\examples{
ddmmssToDecimal(38, 37, 38, TRUE) # latitude of St. Louis, Missouri, USA
ddmmssToDecimal(90, 11, 52.1, FALSE) # longitude of St. Louis, Missouri, USA
}
