% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalToDms.r
\name{decimalToDms}
\alias{decimalToDms}
\title{Convert geographic coordinates in decimal format to degrees-minutes-second}
\usage{
decimalToDms(x, hemis)
}
\arguments{
\item{x}{Numeric, longitude or latitude in decimal format.}
}
\value{
A numeric vector with three named elements: degrees, seconds, and seconds.
}
\description{
This function converts geographic coordinates in decimal format to degrees-minutes-seconds (DD-MM-SS) format.
}
\examples{
decimalToDms(38.56123) # latitude of St. Louis, Missouri, USA
decimalToDms(38.56123) # latitude of St. Louis, Missouri, USA
decimalToDms(90.06521) # longitude of St. Louis, Missouri, USA
decimalToDms(90.06521) # longitude of St. Louis, Missouri, USA
}
