% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmsToDecimal.r
\name{dmsToDecimal}
\alias{dmsToDecimal}
\title{Convert geographic coordinates in degrees-minutes-second to decimal format}
\usage{
dmsToDecimal(dd, mm, ss, hemis)
}
\arguments{
\item{dd}{Numeric. Degrees longitude or latitude. Can be a decimal value.}

\item{mm}{Numeric. Minutes longitude or latitude. Can be a decimal value.}

\item{ss}{Numeric. Second longitude or latitude. Can be a decimal value.}

\item{hemis}{Logical or character. If logical, set equal to \code{TRUE} if the location is east of the Prime Meridian (longitude) or north of the equator (latitude). If a character, this can be "N" (north), "S" (south), "E" (east), or "W" (west).}
}
\value{
Numeric.
}
\description{
This function converts geographic coordinates in degrees-minutes-seconds (DD-MM-SS) format to decimal format.
}
\examples{
dmsToDecimal(38, 37, 38, TRUE) # latitude of St. Louis, Missouri, USA
dmsToDecimal(38, 37, 38, 'N') # latitude of St. Louis, Missouri, USA
dmsToDecimal(90, 11, 52.1, FALSE) # longitude of St. Louis, Missouri, USA
dmsToDecimal(90, 11, 52.1, 'W') # longitude of St. Louis, Missouri, USA
}
