% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_calcTol.r
\name{.calcTol}
\alias{.calcTol}
\title{Automatically calculate tolerance for null model point randomization}
\usage{
.calcTol(dists, removeTopRow)
}
\arguments{
\item{dists}{Either a lower-triangular square matrix (with the diagonal set to \code{NA}) or a rectangular matrix. Values represent pairwise distances between points.}

\item{removeTopRow}{Logical, if \code{TRUE} then remove the top row of the \code{dists} matrix before calculating summary statistics. If the matrix is lower-triangular and the diagonal is \code{NA} then retaining the top row will cause a warning. This value should be \code{FALSE} if \code{dists} is a rectangular matrix or otherwise not triangular!}
}
\value{
Numeric.
}
\description{
This function takes a triangular or rectangular distance matrix representing pairwise distances between points and calculates a tolerance value for spatial randomizations of points meant to reflect the spatial autocorrelation inherent in the observed points. It is meant as a helper function for the family of \code{randPoints---} functions.
}
\details{
Tolerance is calculated as half of the value of minimum pairwise distances that are >0.
}
\keyword{internal}
