% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioticVelocity.r
\name{.cardinalDistance}
\alias{.cardinalDistance}
\title{Movement of occupied cells in a given direction of a fixed point}
\usage{
.cardinalDistance(
  direction,
  refLong,
  refLat,
  x1,
  x2,
  x1weightedLongs,
  x1weightedLats,
  x2weightedLongs,
  x2weightedLats,
  longOrLat
)
}
\arguments{
\item{refLong}{Numeric, longitude of reference point from which to partition the weights into a northern, southern, eastern, or western portion.}

\item{refLat}{Numeric, latitude of reference point.}

\item{x1}{Matrix of weights in time 1 (i.e., population size).}

\item{x2}{Matrix of weights in time 2 (i.e., population size).}

\item{x1weightedLongs}{Matrix of longitudes weighted (i.e., by population size, given by \code{x1}).}

\item{x1weightedLats}{Matrix of latitudes weighted (i.e., by population size, given by \code{x1}).}

\item{x2weightedLongs}{Matrix of longitudes weighted (i.e., by population size, given by \code{x2}).}

\item{x2weightedLats}{Matrix of latitudes weighted (i.e., by population size, given by \code{x2}).}

\item{longOrLat}{Numeric matrix, latitude or longitudes. If \code{direction} is \code{'n'} or \code{'s'} this must be latitudes. If \code{direction} is \code{'e'} or \code{'w'} this must be longitudes.}
}
\value{
a list object with distance moved and abundance of all cells north/south/east/west of reference point.
}
\description{
This function calculates the weighted distance moved by a mass represented by set of cells which fall north, south, east, or west of a given location (i.e., typically the centroid of the starting population). Values >0 confer movement to the north, south, east, or west of this location. #' @param direction Any of: \code{'n'} (north), \code{'s'} (south), \code{'e'} (east), or \code{'w'} (west).
}
\keyword{internal}
