% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_cardinalVelocity.r
\name{.cardinalVelocity}
\alias{.cardinalVelocity}
\title{Movement of occupied cells in a given direction of a fixed point}
\usage{
.cardinalVelocity(
  direction,
  x2,
  x2weightedLongs,
  x2weightedLats,
  refLong,
  refLat,
  longOrLat
)
}
\arguments{
\item{direction}{Any of: \code{'n'} (north), \code{'s'} (south), \code{'e'} (east), or \code{'w'} (west).}

\item{x2}{Matrix of weights (i.e., population size).}

\item{x2weightedLongs}{Matrix of longitudes weighted (i.e., by population size, given by \code{x2}).}

\item{x2weightedLats}{Matrix of latitudes weighted (i.e., by population size, given by \code{x2}).}

\item{refLong}{Numeric, longitude of reference point.}

\item{refLat}{Numeric, latitude of reference point.}

\item{longOrLat}{Numeric matrix, latitude or longitudes. If \code{direction} is \code{'n'} or \code{'s'} this must be latitudes. If \code{direction} is \code{'e'} or \code{'w'} this must be longitudes.}
}
\value{
a list object with distance moved and abundance of all cells north/south/east/west of reference point.
}
\description{
This function calculates the weighted distance moved by a mass represented by set of cells which fall north, south, east, or west of a given location (i.e., typically the centroid of the starting population). Values >0 confer movement to the north, south, east, or west of this location. Values equal to 0 represent no cells with mass different from 0 were north, south, east, or west of the reference location.
}
\keyword{internal}
