% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_euclid.r
\name{.euclid}
\alias{.euclid}
\title{Euclidean distance between a pair of points}
\usage{
.euclid(x1, y1, x2 = NULL, y2 = NULL)
}
\arguments{
\item{x1}{Numeric}

\item{y1}{Numeric}

\item{x2}{Numeric}

\item{y2}{Numeric}
}
\description{
Euclidean distance between a pair of points or two points.
}
\details{
If \code{x2} and \code{y2} are \code{NULL} then the output is simply \code{x1 - y2}.
}
\keyword{internal}
