% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioticVelocity.r
\name{.interpolateLongFromMatrix}
\alias{.interpolateLongFromMatrix}
\title{Longitude of of a quantile of the geographic abundance distribution-}
\usage{
.interpolateLongFromMatrix(prob, x, longitude)
}
\arguments{
\item{prob}{Quantile value (i.e., in the range [0, 1])}

\item{x}{Matrix of abundances.}

\item{longitude}{Matrix of latitudes.}
}
\description{
This function returns the longitude of a quantile of the geographic abundance distribution. The input is derived from a rasterized map of the species abundance distribution. If a quantile would occur somewhere between two columns, the longitude is linearly interpolated between the latitudes of the two columns bracketing its value. It will probably return \code{NA} if the quantile falls outside the range of the values.
}
\keyword{internal}
