% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_randPointsReturn.r
\name{.randPointsReturn}
\alias{.randPointsReturn}
\title{Return output common to "randPointsRespecting~" functions}
\usage{
.randPointsReturn(x, randSites, crs, keepData, out = NULL)
}
\arguments{
\item{x}{Original input data (matrix, data frame, SpatialPoints, or SpatialPointsDataFrame).}

\item{randSites}{SpatialPoint object.}

\item{crs}{Character or object of class \code{CRS}.}

\item{keepData}{Logical, if \code{TRUE} then the original data in \code{x} (i.e., columns that do not represent coordinates) will be retained in the output but the coordinates will be shuffled. If \code{FALSE} (default) then the returned value will have just shuffled coordinates.}

\item{out}{\code{NULL} (should be \code{NULL} if \code{keepData} is \code{FALSE}), \emph{or} a copy of the original input data (the unshuffled \code{x}).}
}
\description{
Returns an object of same class as input unless user desired to strip non-geographic data from the output.
}
\keyword{internal}
