% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_sampleRastSubregion.r
\name{.sampleRastSubregion}
\alias{.sampleRastSubregion}
\title{Select random points from a portion of a raster}
\usage{
.sampleRastSubregion(rast, n, width, restrict = TRUE, circleCent = NULL, ...)
}
\arguments{
\item{rast}{Raster, RasterStack, or RasterBrick demarcating the area in which randomized sites are to be placed. If a RasterStack or a RasterBrick is used then the first layer will be used (i.e., so cells with \code{NA} will not have points located within them).}

\item{n}{Positive integer, number of random points to select.}

\item{width}{Positive numeric, size of buffer radius (in map units--usually meters).}

\item{restrict}{Logical, if \code{TRUE} (default), then select random points from any non-\code{NA} cell in a restricted area. If \code{FALSE}, then select from any non-\code{NA} cell on the raster. If \code{FALSE}, then is exactly the same as using \code{\link[raster]{randomPoints}} exept that it automatically returns a SpatialPoints object for the random points.}

\item{circleCent}{A SpatialPoints or SpatialPointsDataFrame object. The default value is \code{NULL}, in which case a centroid for the circle is randomly chosen. If supplied, then this will be the center of the circle.}

\item{...}{Other arguments (not used).}
}
\value{
A list object with four elements: 1) A SpatialPoints object of randomly selected points; 2) a SpatialPoints object representing the  center of the circle chosen to represent the subregion; 3) a numeric value representing the buffer width; and 4) a numeric value representing the number of non-\code{NA} cells in the region from which the points were selected.
}
\description{
This function selects random points from a randomly-selected subregion of raster. The subregion is selected by randomly locating a point on the any non-\code{NA} cell in the raster, buffering it by a given width, rasterizing the buffered area, then selecting from any non-\code{NA} cells in this region.
}
\keyword{internal}
