% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmSdm.r
\docType{package}
\name{enmSdm}
\alias{enmSdm}
\alias{enmSdm-package}
\title{enmSdm: Species distribution modeling and ecological niche modeling}
\description{
This package contains tools for modeling the distributions and niches of species or species-like entities. Its main features are a set of functions for training ENMs, evaluating niche overlap, and correcting for sampling bias.
}
\details{
Create an issue on \href{https://github.com/adamlilith/enmSdm/issues}{GitHub}.
}
\author{
Adam B. Smith
}
