% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLarge.r
\name{extractLarge}
\alias{extractLarge}
\title{Memory-wise extraction of data from a raster or spatial object}
\usage{
extractLarge(x, y, keep = NULL, atATime = 10000, verbose = TRUE, ...)
}
\arguments{
\item{x}{\code{SpatRaster} or \code{SpatVector} from which data will be out.}

\item{y}{\code{SpatVector} with points/lines/polygons denoting locations where data will be out.}

\item{keep}{Names or indices of the fields in \code{x} to keep. If \code{NULL} (default), then all columns will be kept and there will be very little advantage to using this function over just using \code{extract}.}

\item{atATime}{Number of records in \code{y} to extract at one time.}

\item{verbose}{Report progress (\code{TRUE}) or not (\code{FALSE}).}

\item{...}{Arguments to pass to \code{\link[terra]{extract}}.}
}
\value{
Data frame.
}
\description{
This function is the same as the \code{\link[terra]{extract}} function in the \pkg{terra} package, except that it is built to better handle cases when the points/lines/polygons where the extractions are are very large in memory. Essentially, the function is a wrapper for \code{extract} with a for loop that allows you to discard unwanted columns. Note that this function is not intended to make things fast... just feasible
}
\examples{

}
