% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpbOneThreshold.r
\name{fpbOneThreshold}
\alias{fpbOneThreshold}
\title{Calculate the Fpb measure of model accuracy for a single threshold}
\usage{
fpbOneThreshold(threshold, pres, bg, presWeight = rep(1, length(pres)),
  bgWeight = rep(1, length(bg)), tr = NULL)
}
\arguments{
\item{threshold}{Numeric value or numeric vector within the range [0, 1]. Threshold value(s) at which to calculate Fpb.}

\item{pres}{Numeric vector. Predictions at presences.}

\item{bg}{Numeric vector. Predictions at background sites.}

\item{presWeight}{Numeric same length as \code{pres}. Weights for presence predictions.}

\item{bgWeight}{Numeric same length as \code{bg}. Weights for background predictions.}

\item{ttr}{Same as \code{threshold} but deprecated. Included for backwards compatability.}
}
\value{
Numeric.
}
\description{
Calculate the Fpb measure of model accuracy for a single threshold
}
