% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCores.r
\name{getCores}
\alias{getCores}
\title{Calculate number of cores to use for multi-core functions}
\usage{
getCores(rast, cores = 2, forceMulti = FALSE)
}
\arguments{
\item{rast}{Raster, RasterStack, or RasterBrick}

\item{cores}{Integer > 0. Number of cores desired to be used. Default is 2.}

\item{forceMulti}{Logical. If \code{TRUE} then force number of cores to be >1 even if the raster processing can be done in one chunk. Default is \code{FALSE}.}
}
\value{
Integer.
}
\description{
This function calculates the number of cores to use for multi-core functions based on user preferences, size of the raster object, number of available cores, and whether or not the user wants to use multi-core functionality even if it is possible to do the entire raster processing in one block.
}
