% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemurs.r
\docType{data}
\name{lemurs}
\alias{lemurs}
\title{Lemur occurrences from GBIF}
\format{
An object of class \code{'data.frame'}.
}
\source{
www.gbif.org
}
\usage{
data(lemurs)
}
\description{
Data frame of lemur occurrences
}
\examples{
data(lemurs)
lemurs
}
\keyword{Madagascar}
\keyword{datasets,}
\keyword{lemurs,}
