% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longLatRasters.r
\name{longLatRasters}
\alias{longLatRasters}
\title{Generate rasters with cell values equal to cell longitude or latitude}
\usage{
longLatRasters(x, m = NULL, filePath = NULL, ...)
}
\arguments{
\item{x}{Raster* object. The output will have the same resolution, extent, and coordinate projection system as \code{x}.}

\item{m}{Logical or raster* object. If TRUE then the output rasters will have \code{NA}s placed in cells where \code{NA}s exist in \code{x} (or the first layer in \code{x} if \code{x} is a raster stack or brick. If \code{FALSE} then longitude and latitude will be calcualted for all cells. If \code{m} is a raster then values in the output will be \code{NA} wherever they are \code{NA} in \code{m}.}

\item{filePath}{String or \code{NULL}. If a string, then this is the path (not including file name) to which to write the raster stack with longitude/latitude rasters. If \code{NULL} then no file is written.}

\item{...}{Arguments to pass to \code{writeRaster} (if \code{filePath} is not \code{NULL}).}
}
\value{
Object of class \code{RasterStack}.
}
\description{
This function generates a raster stack with two rasters, one with cell values equal to the cell's longitude and the other with cell values equal to cthe cell's latitude.
}
\examples{
\donttest{
x <- raster::raster() # raster with 1 deg resolution and extent equal to entire world
ll <- longLatRasters(x)
plot(ll)
}
}
