% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad0.r
\docType{data}
\name{mad0}
\alias{mad0}
\title{Magascar spatial object}
\format{
An object of class \code{'SpatialPolygonsDataFrame'}.
}
\source{
www.gadm.org
}
\usage{
data(mad0)
}
\description{
SpatialPolygonsDataFrame of Madagascar from GADM
}
\examples{
data(mad0)
mad0
plot(mad0)
}
\keyword{Madagascar}
\keyword{datasets,}
