% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad1.r
\docType{data}
\name{mad1}
\alias{mad1}
\title{Madagascar spatial object}
\format{
An object of class \code{'SpatialPolygonsDataFrame'}.
}
\source{
www.gadm.org
}
\usage{
data(mad0)
}
\description{
SpatialPolygonsDataFrame of Madagascar from GADM
}
\examples{
data(mad1)
mad1
plot(mad1)
}
\keyword{Madagascar}
\keyword{datasets,}
