% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLarsData.r
\name{makeLarsData}
\alias{makeLarsData}
\title{Add columns to a data matrix to represent polynomial and interaction terms}
\usage{
makeLarsData(data, resp, preds, scale = TRUE, quadratic = TRUE,
  cubic = TRUE, interaction = TRUE, interQuad = TRUE,
  na.rm = FALSE)
}
\arguments{
\item{data}{Data frame.}

\item{resp}{Character or integer or \code{NULL}. Name or index of colum in \code{data} that represents the response variable. If \code{NULL} then it is assumed there is no response column in \code{data}.}

\item{preds}{Character or integer. Names or indices of columns to use as predictors.}

\item{scale}{Logical \emph{or} a list. If TRUE then scale values in \code{data[ , preds]} to have mean of 0 and unit variance. Note that scaling is done before adding terms. If a list, then this is the same as, for example, the two attributes from \code{attributes(scale(data[ , preds]))} named \code{`scaled:center`}and \code{`scaled:scale`}. Ergo, they each are list of the centers and scales (usually means and standard deviation) of each column of \code{data[ , preds]}, and each has names given by \code{preds}.}

\item{quadratic}{Logical. If TRUE then include quadratic terms in model construction stage for non-factor predictors. Quadratic columns will be named \code{<predictor name>_pow2}.}

\item{cubic}{Logical. If TRUE then include cubic terms in model construction stage for non-factor predictors. Cubic columns will be named \code{<predictor name>_pow3}.}

\item{interaction}{Logical. If TRUE then include 2-way interaction terms (including interactions between factor predictors). Interaction columns will be named \code{<predictor 1 name>_by_<predictor 2 name>}.}

\item{interQuad}{Logical. If TRUE then include all possible interactions of the form \code{x * y^2} unless \code{y} is a factor (linear-by-quadratic features). Linear-by-quadratic columns will be named \code{<predictor 1 name>_by_<predictor 2 name>_pow2}.}

\item{na.rm}{Logical. If TRUE then remove all rows of \code{data} in which there is at least one \code{NA} among \code{resp} or \code{preds}. The default is FALSE, which will cause an error if any row has an \code{NA}.}
}
\value{
An object of class \code{larsData} (which is also a \code{list}) with six elements:
* A character named "\code{resp}" indicating the name of the column of that contains the response variable.
* A character list named "\code{preds}" indicating the name of the columns of that contain the original predictors.
* A data frame named "\code{data}" containing \code{data} but with extra columns representing the added terms;
* A list object named "\code{scales}" representing the scale (mean and standard derviation) used to center and rescale values in the data frame; and
* A list named "\code{groups}" with groups (names of predictors that should be considered together based on marginality).
* A list named "\code{features}" indicating what kind of features were added to \code{data} (e.g., quadratic, cubic, etc.).
}
\description{
This function adds columns to a data matrix representing quadratic, cubic, 2-way-interaction, and linear:quadratic interactions. It is especially useful for preparing a data matrix for \code{\link{trainLars}} or \code{\link{predictLars}}.
}
\details{
If \code{scale} is \code{TRUE} then predictors with zero variance will be removed from the data before creating higher-order terms.
}
\examples{
set.seed(123)
x <- data.frame(y=c(rep(1, 10), rep(0, 10)), x1=1:10, x2=runif(20) * 1:20, x3=rnorm(20) - 1:20)
out <- makeLarsData(x, resp='y', preds=c('x1', 'x2', 'x3'))
str(out)
}
\seealso{
\code{\link{trainLars}}, \code{\link{predictLars}}
}
