% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLars.r
\name{predictLars}
\alias{predictLars}
\title{Predict using Least Angle Regression (LARS) model}
\usage{
predictLars(object, newdata, type = c("link", "response"), preds = NULL,
  failOnMissing = FALSE, ...)
}
\arguments{
\item{object}{Object of classes \code{larsModel} and \code{cv.grpregOverlapMulti}.}

\item{newdata}{Data frame.}

\item{type}{Character. Type of prediction to make: \code{link} produces predictions on the scale of the predictors, \code{response} produces predictions on the scale of the response (i.e., in the range [0, 1] if \code{family = 'binomial'} in \code{trainLars}).}

\item{preds}{\code{NULL} or a character list. If \code{NULL} then all predictors (and their higher-order terms) will be used in the prediction. If a character vector that is a subset of the predictors used in training the LARS model, then the function yields partial predictions using \emph{only} terms that include the specified variable(s) (plus any quadratic, cubic, interaction, and linear-by-quadratic terms, if the model was trained using them).
failOnMissing Logical. If \code{TRUE} then if no variable(s) stated in \code{pred} occurs in the model then cause an error. Variables can be missing either because the user did not include them in the original model or because the variable(s) had no variability and so were removed automatically in the model training process. Note that if \code{failOnMissing} is \code{TRUE} and there is at least one variable in \code{preds} that occurs in the model then this function will still return values (it will not cause an error). If \code{failOnMissing} is FALSE and no variables in \code{preds} occurs in the model then a warning will be produced and the returned value will be \code{NULL}.}

\item{...}{Arguments to pass to \code{\link[grpregOverlap]{predict.grpregOverlap}}.}
}
\value{
Numeric.
}
\description{
This function makes calculates predictions from a \code{grpregOverlap} object constructed using the \code{\link{trainLars}} function..
}
\seealso{
\code{\link[grpreg]{grpreg}}, \code{\link[grpreg]{predict.grpreg}}, \code{\link[grpregOverlap]{grpregOverlap}}, \code{\link[grpregOverlap]{predict.grpregOverlap}}, \code{\link{makeLarsData}}, \code{\link{trainLars}}
}
