% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMaxNet.r
\name{predictMaxNet}
\alias{predictMaxNet}
\title{Predictions from a MaxNet model}
\usage{
predictMaxNet(model, newdata, clamp = TRUE, type = "cloglog", ...)
}
\arguments{
\item{model}{Object of class \code{maxnet}.}

\item{newdata}{Object of class \code{data.frame}, \code{SpatRaster} (\pkg{terra} package), or any of \code{Raster}, \code{RasterBrick}, or \code{RasterStack} (\pkg{raster} package).}

\item{clamp}{If \code{TRUE} (default), predict outside the range of training data by 'clamping' values to the last value.}

\item{type}{One of:
\itemize{
    \item        \code{cloglog} (default): Predictions are on a complementary log-log scale.
    \item        \code{logistic}: Predictions are on a logistic scale (and thus technically the same to several decimal places as predictions from MaxEnt <=3.3.3k, except for differences in default features).
    \item        \code{link}: Predictions are on the scale of the predictors.
    \item        \code{exponential}: Predictions are on an exponential ('raw') scale.
}}

\item{...}{Other arguments (unused).}
}
\value{
Numeric vector.
}
\description{
This function is the same as the \code{predict} function in the \pkg{maxnet} package, except that:
\itemize{
	\item	If the input is a data frame, the output is a vector as output (not a single-column matrix);
	\item	If the input is a \code{SpatRaster} or \code{Raster*}, the output is of the same type;
	\item	The default output is on the cloglog scale
	\item   The function can be explicitly called (versus doing, say, \code{maxnet:::predict.maxnet}, which does not work even when that would be really useful...).
}
}
\seealso{
\code{\link[raster]{predict}} from the \pkg{raster} package, \code{\link[terra]{predict}} from the \pkg{terra} package, and \code{\link[maxnet]{maxnet}} (see the \code{predict} function therein)
}
