% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randGeoBySelfAndOther.r
\name{randGeoBySelfAndOther}
\alias{randGeoBySelfAndOther}
\title{Randomizes the location of two sets of geographic points while respecting spatial autocorrelation}
\usage{
randGeoBySelfAndOther(x1, x2, rast, bins = 20, tol = 0.001,
  distFunct = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x1}{Matrix, data frame, SpatialPoints, or SpatialPointsDataFrame object. If this is a matrix or data frame, the first two columns must represent longitude and latitude (in that order). If \code{x} is a matrix or data frame, the coordinates are assumed to be unprojected (WGS84) (a coordinate reference system proj4 string or \code{CRS} object can be passed into the function using \code{...}). If \code{x} is a SpatialPoints or SpatialPointsDataFrame and not in WGS84 or NAD83, then coordinates are projected to WGS84 (with a warning).}

\item{x2}{As \code{x1}.}

\item{rast}{Raster, RasterStack, or RasterBrick used to locate presences randomly. If this is a RasterStack or a RasterBrick then the first layer will be used (i.e., so cells with \code{NA} will not have points located within them).}

\item{bins}{Integer > 1, number of overlapping bins across which to calculate distribution of pairwise distances between points. The range covered by bins starts at 0 and and at the largest observed pairwise distance + 0.5 * bin width. Default value is 20.}

\item{tol}{Numeric >0, for any one bin, root-mean square deviation between observed pairwise distribution of distances and randomized distances required for the randomized distances to be considered statistically the same as observed distances.}

\item{distFunct}{Either a function or \code{NULL}. If \code{NULL} then \code{\link[geosphere]{distCosine}} is used to calculate distances.  Other "dist" functions (e.g., \code{\link[geosphere]{distGeo}}) can be used.  Alternatively, a custom function can be used so long as its first argument is a 2-column numeric matrix with one row for the x- and y-coordinates of a single point and its second argument is a two-column numeric matrix with one or more rows of other points.}

\item{verbose}{Logical, if \code{FALSE} (default) show no progress indicator. If \code{TRUE} then display occasional updates and graph.}

\item{...}{Arguments to pass to \code{distCosine} or \code{\link[dismo]{randomPoints}}. Note that if \code{x} is a matrix or data frame a coordinate reference system may be passed using \code{crs = <proj4 string code} or \code{crs = <object of class CRS (see sp package)>}. Otherwise WGS84 is assumed.}
}
\value{
Object of the same class as \code{x} but with coordinates randomized.
}
\description{
This function randomizes the location of two sets of geographic points with respect to one another retaining (more or less) the same distribution of pairwise distances between points with and between sets (plus or minus a user-defined tolerance).
}
\examples{
# madagascar
library(dismo)
madElev <- getData('alt', country='MDG')
par(layout(matrix(c(1, 2), nrow=1)))
plot(madElev, main='Madagascar')
data(lemur)
points(lemur, pch=16)
rands <- randGeoBySelf(lemur, mad, verbose=TRUE)
par(fig=1, new=FALSE)
points(rand, col='red')
}
\seealso{
\code{\link[dismo]{randomPoints}}, \code{\link[enmSdm]{randGeoBySelf}}
}
