% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPointsBatchExtract.r
\name{randPointsBatchExtract}
\alias{randPointsBatchExtract}
\title{Updates a list of random points with environmental data from rasters}
\usage{
randPointsBatchExtract(rands, rast, verbose = FALSE)
}
\arguments{
\item{rands}{A list object typically generated using the \code{\link[enmSdm]{randPontsMaster}} function.}

\item{rast}{A raster, raster stack, or raster brick from which to extract data.}

\item{verbose}{Logical, if \code{TRUE} display progress.}
}
\value{
A list.
}
\description{
This function is called using a list object typically generated using the \code{\link[enmSdm]{randPointsMaster}} function. To each set of random points represented in that list it adds environmental data extracted from a raster stack or brick.
}
\seealso{
\code{\link[enmSdm]{randPointsRespectingSelf}}, \code{\link[enmSdm]{randPointsRespectingSelfOther1}}, \code{\link[enmSdm]{randPointsRespectingSelfOther2}}, \code{\link[enmSdm]{randPointsBatch}}, \code{\link[enmSdm]{randPointsBatchSampled}}, \code{\link[enmSdm]{randPointsBatchNicheOverlap}}
}
