% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPointsBatchNicheOverlap.r
\name{randPointsBatchNicheOverlap}
\alias{randPointsBatchNicheOverlap}
\title{Niche overlap for a set of iterated "randPointsRespecting~" functions}
\usage{
randPointsBatchNicheOverlap(rands, env, vars, x = NULL, bins = 100,
  cor = TRUE)
}
\arguments{
\item{rands}{A list object typically generated using the \code{\link[enmSdm]{randPontsMaster}} function.}

\item{env}{Either a data frame, matrix, or any object that can be coerced to a data frame containing environmental data at available background sites, \emph{or} an object of class \code{princomp} representing a principal components analysis generated using the \code{\link[stats]{princomp}} function with argument \code{scores = TRUE}.}

\item{vars}{Either a character list naming columns in \code{x1}, \code{x2}, and \code{x3} to be used as environmental data, \emph{or} positive integers indexing the columns to be used as environmental data.}

\item{x}{Either \code{NULL} (default) or a data frame, matrix, SpatialPointsDataFrame, or other object that can be coerced to a data frame. If supplied then the objects (usually species) represented in \code{rands} are compared to this set of environmental values. This argument \emph{must} be supplied if \code{rands} was generated using \code{\link[enmSdm]{randPointsRespectingSelf}} or \code{\link[enmSdm]{randPointsRespectingSelfOther1}}.}

\item{bins}{Number of bins into which to divide the environmental space (default is 100 on each side).}

\item{cor}{Logical, if \code{TRUE} (default), then the PCA used to construct the environmental space will use the correlation matrix (this is highly recommended if the variables are on different scales). This is ignored if \code{env} is an object of class \code{princomp}.}
}
\value{
Data frame.
}
\description{
This function is called using a list object typically generated using the \code{\link[enmSdm]{randPontsMaster}} function (plus sometimes followed by the \code{\link[enmSdm]{randPointsExtract}} and \code{\link[enmSdm]{randPointsSampled}} functions). It calculates niche overlap for each set of points. Essentially it is a wrapper for \code{\link[enmSdm]{nicheOverlap}}.
}
\seealso{
\code{\link[enmSdm]{randPointsRespectingSelf}}, \code{\link[enmSdm]{randPointsRespectingSelfOther1}}, \code{\link[enmSdm]{randPointsRespectingSelfOther2}}, \code{\link[enmSdm]{randPointsBatch}}, \code{\link[enmSdm]{randPointsBatchSampled}}, \code{\link[enmSdm]{randPointsBatchExtract}}
}
