% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPointsBatchSampled.r
\name{randPointsBatchSampled}
\alias{randPointsBatchSampled}
\title{Returns collated set of sampled sites from a set of "randPoints~" calls}
\usage{
randPointsBatchSampled(rands)
}
\value{
A \code{\link[sp]{SpatialPoints}} object.
}
\description{
This function is called using a list object typically generated using the \code{\link[enmSdm]{randPontsMaster}} function. It returns a \code{SpatialPoints} object that represents all of the randomized points from all of the iterations. These can be used as the "available environment" in a niche overlap test.
}
\seealso{
\code{\link[enmSdm]{randPointsRespectingSelf}}, \code{\link[enmSdm]{randPointsRespectingSelfOther1}}, \code{\link[enmSdm]{randPointsRespectingSelfOther2}}, \code{\link[enmSdm]{randPointsBatch}}, \code{\link[enmSdm]{randPointsBatchExtract}}, \code{\link[enmSdm]{randPointsBatchNicheOverlap}}
}
