% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randPointsRespectingSelf.r
\name{randPointsRespectingSelf}
\alias{randPointsRespectingSelf}
\title{Randomizes geographic points while observing spatial autocorrelation}
\usage{
randPointsRespectingSelf(x, rast, tol = NULL, w = NULL,
  distFunct = NULL, restrict = TRUE, keepData = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{x}{Matrix, data frame, SpatialPoints, or SpatialPointsDataFrame object. If this is a matrix or data frame, the first two columns must represent longitude and latitude (in that order). If \code{x} is a matrix or data frame, the coordinates are assumed to be unprojected (WGS84) (a coordinate reference system proj4 string or \code{CRS} object can be passed into the function using \code{...}). If \code{x} is a SpatialPoints or SpatialPointsDataFrame and not in WGS84 or NAD83, then coordinates are projected to WGS84 (with a warning).}

\item{rast}{Raster, RasterStack, or RasterBrick demarcating the area in which randomized sites are to be placed. If a RasterStack or a RasterBrick is used then the first layer will be used (i.e., so cells with \code{NA} will not have points located within them).}

\item{tol}{Numeric >0, maximum root-mean-square distance allowed between the set of observed pairwise distances between points and the set of randomized pairwise distances between points. The algorithm will shuffle points until the calculated difference is <= this number. Units are the same as units used by the coordinate reference system of \code{x} (usually meters).}

\item{w}{Either \code{NULL} (default), a numeric matrix, or a function. This argument can be used to specify a set of weights to apply to the root-mean-square distance matrix. It must either be \code{NULL}, in which case all squared distances have equal weight, a symmetrical matrix with the same number of rows and columns as the length of \code{x} (diagonal elements are ignored), or a function. If \code{w} is a matrix then they are applied as \code{sqrt(sum(w * (d_obs - d_rand)^2) / sum(w))} where \code{d_obs} is the observed pairwise distance between points and \code{d_rand} the pairwise distance between the randomized points. If it is a function it is applied as \code{sqrt(mean(w(d_obs - d_rand)^2))}. Note that any weights other than \code{NULL} will also affect the desirable value of \code{tol}. This argument is especially useful for determining the scale of structure desired to be replicated in the randomized data. For example, using an increasing function will cause the randomization to recreate broad-scale patterns in the data (e.g., general shape and extent of \code{x}) but allow fine-scale patterns to be more random. A decreasing function will allow the broad pattern of \code{x} to shift (e.g., "bend") but retain fine-scale pattern.}

\item{distFunct}{Either a function or \code{NULL}. If \code{NULL} then \code{\link[geosphere]{distGeo}} is used to calculate distances. Other "dist" functions (e.g., \code{\link[geosphere]{distGeo}}) can be used for lower/increased accuracy and faster/slower performance. Alternatively, a custom function can be used so long as its input and output is comparable to any of these functions.}

\item{restrict}{Logical, if \code{TRUE} (default), then select random points from any non-\code{NA} cell in a restricted area. If \code{FALSE}, then select from any non-\code{NA} cell on the raster. If \code{FALSE}, then random sites are selected from across the entire raster.}

\item{keepData}{Logical, if \code{TRUE} then the original data in \code{x} (i.e., columns that do not represent coordinates) will be retained in the output but the coordinates will be shuffled. If \code{FALSE} (default) then the returned value will have just shuffled coordinates.}

\item{verbose}{Logical, if \code{FALSE} (default) show no progress indicator. If \code{TRUE} then display progress.}

\item{...}{Arguments to pass to \code{distGeo} or \code{\link[enmSdm]{sampleRast}}. Note that if \code{x} is a matrix or data frame a coordinate reference system may be passed using \code{crs = <proj4 string code>} or \code{crs = <object of class CRS>} (see \pkg{sp} package). Otherwise the coordinates are assumed to be unprojected (WGS84).}
}
\value{
Object of the same class as \code{x} but with coordinates randomized.
}
\description{
This function randomizes the location of geographic points while retaining (more or less) the same distribution of pairwise distances between points (plus or minus a user-defined tolerance). The procedure is meant to generalize the "RTR" (rotate/translate/reflect) randomization procedure proposed by Nunes, L.A. and Pearson, R.G. 2017. A null biogeographical test for assessing ecological niche evolution. \emph{Journal of Biogeography} 44:1331-1343. The algorithm is a conceptual adaptation of the randomization procedure presented in Beale, C.M., J.J. Lennon, and A. Gimona. 2008. Opening the climate envelope reveals no macroscale associations with climate in European birds. \emph{Proceedings of the National Academy of Sciences USA} 105:14908-14912.
}
\details{
The argument \code{startBy} can be used to speed convergence to a solution. The two options, \code{anywhere} and \code{restricted}, start with patterns that are very over-dispersed or somewhat over-dispersed, respectively, but the \code{restricted} option will be slower to start. Under this option start up is initiated by placing a single random point on the landscape. Then, a buffer with a width equal to 1.05 * half the average maximum pairwise distance is drawn around the point. The raster in \code{rast} is cropped to the extent of this buffer and rasterized to form a circle. Note that if the centroid of the circle lies toward the edge of the raster or if the raster contains \code{NA} values then the circle may be cropped or may otherwise have "holes". Then, a number of random points equal to the number represented by \code{x} are placed within this area. The procedure then starts using these points. Note that the restricted start-up area is only used for the initial placement of points--subsequent randomizations use the entire study region.
}
\seealso{
\code{\link[enmSdm]{randPointsRespectingSelfOther1}}, \code{\link[enmSdm]{randPointsRespectingSelfOther2}}, \code{\link[enmSdm]{randPointsBatch}}, \code{\link[enmSdm]{randPointsBatchSampled}}, \code{\link[enmSdm]{randPointsBatchExtract}}, \code{\link[enmSdm]{randPointsBatchNicheOverlap}}
}
