% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRast.r
\name{sampleRast}
\alias{sampleRast}
\title{Sample random points from a raster with/out replacement.}
\usage{
sampleRast(x, n, adjArea = TRUE, replace = TRUE, prob = TRUE)
}
\arguments{
\item{x}{Raster* object.}

\item{n}{Positive integer. Number of points to draw.}

\item{adjArea}{Logical. If TRUE then adjust probabilities so sampling accounts for cell area.}

\item{replace}{Logical. If TRUE then sample with replacement.}

\item{prob}{Logical. If TRUE then sample cells with probabilities proportional to cell values. If `adjArea` is also TRUE then probabilities are drawn proportional to the product of cell area * the value of the cell.}
}
\value{
2-column matrix with longitude and latitude of random points.
}
\description{
This function returns coordinates randomly located on a raster where cells can be sampled more than once if desired (sampled with replacement) and where the probability of selection is proportinate to the cell value (plus maybe cell area--both if desired).
}
\seealso{
\code{\link[dismo]{randomPoints}}, \code{\link{sampleRastStrat}}
}
