% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRastStrat.r
\name{sampleRastStrat}
\alias{sampleRastStrat}
\title{Sample random points in a stratfied manner from a raster with/out replacement}
\usage{
sampleRastStrat(x, n, adjArea = TRUE, replace = TRUE)
}
\arguments{
\item{x}{Raster* object.}

\item{n}{number of points to sample from *each* stratum}

\item{adjArea}{Logical. If TRUE then adjust sample probability of cells by cell area.}

\item{replace}{Logical. If TRUE then sample with replacement.}
}
\value{
2-column matrix with longitude and latitude of random points.
}
\description{
This function is a robust version of \code{sampleStratified()} in the \code{raster} package where sampling can be performed with replacement in each stratum. Each unique value in a raster is identified as a stratum. Each stratum is sampled a given number of times.
}
\seealso{
\code{\link[raster]{sampleStratified}}, \code{\link[randomPoints]{randomPoints}}, \code{\link{sampleRast}}
}
