% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediWeighted.r
\name{sediWeighted}
\alias{sediWeighted}
\title{Symmetric extremal dependence index (SEDI)}
\usage{
sediWeighted(pres, contrast, presWeight = rep(1, length(pres)),
  contrastWeight = rep(1, length(contrast)), thresholds = seq(0, 1, by
  = 0.01), delta = 0.001, na.rm = FALSE, bg = NULL,
  bgWeight = NULL, ...)
}
\arguments{
\item{pres}{Numeric vector. Predicted values at presence sites.}

\item{contrast}{Numeric vector. Predicted values at absence/background sites.}

\item{presWeight}{Numeric vector same length as \code{pres}. Relative weights of presence sites. The default is to assign each presence a weight of 1.}

\item{contrastWeight}{Numeric vector same length as \code{contrast}. Relative weights of background sites. The default is to assign each presence a weight of 1.}

\item{thresholds}{Numeric vector, Values at which to threshold predictions for calculation of SEDI.}

\item{delta}{Positive numeric >0 in the range [0, 1] and usually very small. This value is used only if calculating the SEDI threshold when any true positive rate or false negative rate is 0 or the false negative rate is 1. Since SEDI uses log(x) and log(1 - x), values of 0 and 1 will produce \code{NA}s. To obviate this, a small amount can be added to rates that equal 0 and subtracted from rates that equal 1.}

\item{na.rm}{Logical. If \code{TRUE} then remove any presences and associated weights and absence/background predictions and associated weights with \code{NA}s.}

\item{bg}{Same as \code{contrast}. Included for backwards compatibility. Ignored if \code{contrast} is not \code{NULL}.}

\item{bgWeight}{Same as \code{contrastWeight}. Included for backwards compatibility. Ignored if \code{contrastWeight} is not \code{NULL}.}

\item{...}{Other arguments (unused).}
}
\value{
Numeric value.
}
\description{
This function calculates the symmetric extremal dependence index (SEDI) using a threshold applied to continuous data to demarcate "presence" from "contrast".
}
\examples{
set.seed(123)
pres <- sqrt(runif(100))
contrast <- runif(10000)
hist(contrast, col='gray', xlim=c(0, 1), breaks=20)
hist(pres, col='green', breaks=20, add=TRUE)
max(sediWeighted(pres, contrast), na.rm=TRUE)
# SEDI is fairly insensitive to weighting (as per Wunderlich et al. 2019)
presWeight <- c(rep(1, 50), rep(1000, 50))
max(sediWeighted(pres, contrast, presWeight=presWeight), na.rm=TRUE)
}
\references{
Wunderlich, R.F., Lin, Y-P., Anthony, J., and Petway, J.R.  2019.  Two alternative evaluation metrics to replace the true skill statistic in the assessment of species distribution models.  Nature Conservation 35:97-116.
}
\seealso{
\code{\link[stats]{cor}}, \code{\link{fpb}}, \code{\link{aucWeighted}}, \code{link[enmSdm]{contBoyce}}, \code{link[enmSdm]{contBoyce2x}}, \code{link[enmSdm]{orssWeighted}}, \code{link[enmSdm]{thresholdWeighted}}, \code{link[enmSdm]{thresholdStats}}
}
