% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensSpecWeighted.r
\name{sensSpecWeighted}
\alias{sensSpecWeighted}
\title{Thresholded evaluation statistics}
\usage{
sensSpecWeighted(thresholds, pres, bg, presWeight = rep(1, length(pres)),
  bgWeight = rep(1, length(bg)), na.rm = FALSE)
}
\arguments{
\item{thresholds}{Numeric or numeric vector. Threshold(s) at which to calculate sensitivity and specificity.}

\item{pres}{Numeric vector. Predicted values at test presences}

\item{bg}{Numeric vector. Predicted values at background/absence sites.}

\item{presWeight}{Numeric vector same length as \code{pres}. Relative weights of presence sites. The default is to assign each presence a weight of 1.}

\item{bgWeight}{Numeric vector same length as \code{bg}. Relative weights of background sites. The default is to assign each presence a weight of 1.}

\item{na.rm}{Logical. If \code{TRUE} then remove any presences and associated weights and background predictions and associated weights with \code{NA}s.}
}
\value{
8-column matrix with the following named columns. \emph{a} = weight of presences >= threshold, \emph{b} = weight of backgrounds >= threshold, \emph{c} = weight of presences < threshold, \emph{d} = weight of backgrounds < threshold, and \emph{N} = sum of presence and background weights.
\itemize{
		\item \code{'threshold'}: Threshold
		\item \code{'sensitivity'}: Sensitivity (\emph{a} / (\emph{a} + \emph{c}))
		\item \code{'specificity'}: Specificity (\emph{d} / (\emph{d} + \emph{b}))
		\item \code{'ccr'}: Correct classification rate ((\emph{a} + \emph{d}) / \emph{N})
		\item \code{'ppp'}: Positive predictive power (\emph{a} / (\emph{a} + \emph{b}))
		\item \code{'npp'}: Negative predictive power (\emph{d} / (\emph{c} + \emph{d}))
		\item \code{'mr'}: Misclassification rate ((\emph{b} + \emph{c}) / \emph{N})
}
}
\description{
This function calculates a series of evaluation statistics based on a threshold or thresholds used to convert continuous predictions to binary predictions.
}
\examples{
set.seed(123)

# set of bad and good predictions at presences
bad <- runif(100)^2
good <- runif(100)^0.1
hist(good, breaks=seq(0, 1, by=0.1), border='green', main='Presences')
hist(bad, breaks=seq(0, 1, by=0.1), border='red', add=TRUE)
pres <- c(bad, good)
bg <- runif(1000)
thresholds <- c(0.1, 0.5, 0.9)
sensSpecWeighted(thresholds, pres, bg)

# upweight bad predictions
presWeight <- c(rep(1, 100), rep(0.1, 100))
sensSpecWeighted(thresholds, pres, bg, presWeight=presWeight)

# upweight good predictions
presWeight <- c(rep(0.1, 100), rep(1, 100))
sensSpecWeighted(thresholds, pres, bg, presWeight=presWeight)
}
\references{
Fielding, A.H. and J.F. Bell. 1997. A review of methods for the assessment of prediction errors in conservation presence/absence models. \emph{Environmental Conservation} 24:38-49.
}
\seealso{
\code{\link[dismo]{threshold}}, \code{\link[enmSdm]{thresholdWeighted}}, \code{\link[dismo]{evaluate}}
}
