% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svToSpatial.r
\name{svToSpatial}
\alias{svToSpatial}
\title{Convert SpatVector to Spatial*}
\usage{
svToSpatial(x)
}
\arguments{
\item{x}{\code{SpatVector} object.}
}
\value{
Object of class \code{Spatial}.
}
\description{
This function converts a \code{SpatVector} object from the \pkg{terra} package to a \code{Spatial} object of the appropriate class (\code{SpatialPoints}, \code{SpatialPointsDataFrame}, \code{SpatialPolygons}, or \code{SpatialPolygonsDataFrame}).
}
\examples{

f <- system.file('ex/lux.shp', package='terra')
v <- terra::vect(f)
spat <- svToSpatial(v)
spat

}
