% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssWeighted.r
\name{tssWeighted}
\alias{tssWeighted}
\title{Weighted True Skill Statistic (TSS)}
\usage{
tssWeighted(pres, bg, presWeight = rep(1, length(pres)),
  bgWeight = rep(1, length(bg)), thresholds = seq(0, 1, by = 0.01),
  na.rm = FALSE)
}
\arguments{
\item{pres}{Numeric vector. Predicted values at test presences}

\item{bg}{Numeric vector. Predicted values at background/absence sites.}

\item{presWeight}{Numeric vector same length as \code{pres}. Relative weights of presence sites. The default is to assign each presence a weight of 1.}

\item{bgWeight}{Numeric vector same length as \code{bg}. Relative weights of background sites. The default is to assign each presence a weight of 1.}

\item{thresholds}{Numeric vector. Thresholds at which to calculate the sum of sensitivity and specificity. The default evaluates all values from 0 to 1 in steps of 0.01.}

\item{na.rm}{Logical. If \code{TRUE} then remove any presences and associated weights and background predictions and associated weights with \code{NA}s.}
}
\value{
Numeric value.
}
\description{
This function calculates the True Skill Statistic (TSS).
}
\details{
This function calculates the maximum value of the True Skill Statistic (i.e., across all thresholds, the values that maximizes sensitivity plus specificity).
}
\examples{
set.seed(123)

# set of bad and good predictions at presences
bad <- runif(100)^2
good <- runif(100)^0.1
hist(good, breaks=seq(0, 1, by=0.1), border='green', main='Presences')
hist(bad, breaks=seq(0, 1, by=0.1), border='red', add=TRUE)
pres <- c(bad, good)
bg <- runif(1000)
tssWeighted(pres, bg)

# upweight bad predictions
presWeight <- c(rep(1, 100), rep(0.1, 100))
tssWeighted(pres, bg, presWeight=presWeight)

# upweight good predictions
presWeight <- c(rep(0.1, 100), rep(1, 100))
tssWeighted(pres, bg, presWeight=presWeight)

e <- dismo::evaluate(pres, bg)
max(e@TPR + e@TNR) - 1

# why different from dismo's evaluate() function?
# because uses different thresholds based on values of
# presence and absence predictions
head(e@t)
}
\references{
See Allouche, O., Tsoar, A., and Kadmon, R. 2006. Assessing the accuracy of species distribution models: Prevalence, kappa and the true skill statistic (TSS). \emph{Journal of Applied Ecology} 43:1223-1232.
}
\seealso{
\code{\link{fpb}}, \code{\link{aucWeighted}}, \code{link[enmSdm]{contBoyce}}, \code{link[enmSdm]{contBoyce2x}}
}
