% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xToCoords.r
\name{xToCoords}
\alias{xToCoords}
\title{Extract geographic coordinates from data frame, matrix, or \code{SpatialPoints}* object}
\usage{
xToCoords(x, longLat = NULL, sp = TRUE)
}
\arguments{
\item{x}{A data frame, matrix, \code{SpatialPoints}, or \code{SpatialPointsDataFrame} object. If a data frame or matrix then the coordinate reference system is assumed to be unprojected (WGS84).}

\item{sp}{Logical. If \code{TRUE} then return object of class\code{SpatialPoints}. If \code{FALSE} then return object of class \code{matrix}.}
}
\value{
Object of class\code{SpatialPoints} or a 2-column \code{matrix}.
}
\description{
This function extracts geographic coordinates from a data frame, matrix, or \code{SpatialPoints}* object
}
\seealso{
\code{\link[sp]{SpatialPoints}}, \code{\link[sp]{SpatialPointsDataFrame}}
}
