% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNiches.r
\name{compareNiches}
\alias{compareNiches}
\title{Metrics of niche overlap}
\usage{
compareNiches(x1, x2, method = c("meanDiff", "meanAbsDiff", "d", "i", "esp",
  "rho", "rankCor"), w = rep(1, length(x1)), na.rm = FALSE, logit = TRUE,
  ...)
}
\arguments{
\item{x1}{Numeric. List of predictions from a model.}

\item{x2}{Numeric. List of predictions from another model.}

\item{method}{Character list, indicates type of metric to calculate:
* \code{meanDiff} mean difference between \code{x1} and \code{x2}
* \code{meanAbsDiff} mean absolute difference between \code{x1} and \code{x2} (ie, \code{sum(abs(x1 - x2))})
* \code{d} Schoener's \emph{D}
* \code{i} Warren's \emph{I}
* \code{esp} Godsoe's \emph{ESP}
* \code{rho} Correlation between \code{x1} and \code{x2} (will apply \code{logitAdj()} first unless logit=FALSE).
* \code{rankCor}  Pearson rank correlation.}

\item{w}{Numeric list. Weights of predictions in \code{x1} and \code{x2}.}

\item{na.rm}{Logical.  If T\code{TRUE} then remove elements in \code{x1} and \code{2} that are \code{NA} in \emph{either} \code{x1} or \code{x2}.}

\item{logit}{Logical. Only used if \code{method} contains \code{rho}.  If TRUE, then \code{x1} and \code{x2} are first adjusted using \code{logitAdj()}. Use \code{...} to alter the default value of \code{epsilon} in \code{logitAdj()}.}

\item{...}{Arguments to send to \code{\link[omnibus]{logitAdj}}.}
}
\value{
List object with one element per value specified by the argument in \code{method}.
}
\description{
This function calculates several metrics of niche overlap based on predictions for two species (or for the same species but different mdels) at the same sites.
}
\examples{
x1 <- seq(0, 1, length.out=100)
x2 <- x1^2
compareNiches(x1, x2)
}
\seealso{
\code{\link{compareResponse}}
}
