% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalToDms.r
\name{decimalToDms}
\alias{decimalToDms}
\title{Convert geographic coordinates in decimal format to degrees-minutes-second}
\usage{
decimalToDms(x)
}
\arguments{
\item{x}{Numeric or vector of numeric values, longitude or latitude in decimal format.}
}
\value{
A numeric matrix with three columns: degrees, seconds, and seconds. Note that the hemisphere (i.e., indicated by the sign of x) is not returned since it could be either north/south or east/west.
}
\description{
This function converts geographic coordinates in decimal format to degrees-minutes-seconds (DD-MM-SS) format.
}
\examples{
decimalToDms(38.56123) # latitude of St. Louis, Missouri, USA
decimalToDms(90.06521) # longitude of St. Louis, Missouri, USA
}
