% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioticVelocity.r
\name{.euclid}
\alias{.euclid}
\title{Euclidean distance between a pair of points}
\usage{
.euclid(a, b)
}
\arguments{
\item{a}{Numeric vector from 1 to 3 elements long}

\item{b}{Numeric vector from 1 to 3 elements long}
}
\description{
Euclidean distance between a pair of points or two points. Note that the output is unsigned if \code{x2} and \code{y2} are provided, but signed if not.
}
\keyword{internal}
