% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_interpolateLongFromMatrix.r
\name{.interpolateLongFromMatrix}
\alias{.interpolateLongFromMatrix}
\title{Longitude of of a quantile of the geographic abundance distribution}
\usage{
.interpolateLongFromMatrix(prob, colsCumSumStd, longitude)
}
\arguments{
\item{prob}{Quantile value (i.e., in the range [0, 1])}

\item{colsCumSumStd}{Vector representing summed rows of a matrix.}

\item{longitude}{Matrix of latitudes.}
}
\description{
This function returns the longitude of a quantile of the geographic abundance distribution. The input is derived from a rasterized map of the species' abundance distribution. If a quantile would occur somewhere between two rows the longitude is linearly interpolated between the latitudes of the two rows bracketing its value. It will probably return \code{NA} if the quantile falls outside the range of the values.
}
\keyword{internal}
