% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheOverlap.r
\name{nicheOverlap}
\alias{nicheOverlap}
\title{Calculate niche overlap as per Broennimann et al. (2012)}
\usage{
nicheOverlap(x1, x2, env, vars, bins = 100, cor = TRUE)
}
\arguments{
\item{x1}{Data frame, matrix, or any object that can be coerced to a data frame containing environmental data at occurrence sites of a species.}

\item{x2}{Data frame, matrix, or any object that can be coerced to a data frame containing environmental data at occurrence sites of another species.}

\item{env}{Either a data frame, matrix, or any object that can be coerced to a data frame containing environmental data at available background sites, \emph{or} an object of class \code{princomp} representing a principal components analysis generated using the \code{\link[stats]{princomp}} function with argument \code{scores = TRUE}.}

\item{vars}{Either a character list naming columns in \code{x1}, \code{x2}, and \code{x3} to be used as environmental data, \emph{or} positive integers indexing the columns to be used as environmental data.}

\item{bins}{Number of bins into which to divide the environmental space (default is 100 on each side).}

\item{cor}{Logical, if \code{TRUE} (default), then the PCA used to construct the environmental space will use the correlation matrix (this is highly recommended if the variables are on different scales). This is ignored if \code{env} is an object of class \code{princomp}.}
}
\value{
List with these named elements:
\itemize{
\item \code{meanDiff} mean difference between binned, standardized densities of \code{x1} and \code{x2} in environmental space.
\item \code{meanAbsDiff} mean absolute difference between binned, standardized densities of \code{x1} and \code{x2} (ie, \code{sum(abs(x1 - x2))}) in environmental space.
\item \code{d} Schoener's \emph{D}.
\item \code{i} Warren's \emph{I}.
\item \code{esp} Godsoe's \emph{ESP}.
\item \code{rho} Correlation between binned, standardized densities of \code{x1} and \code{x2} in environmental space.
\item \code{rankCor}  Pearson rank correlation between binned, standardized densities of \code{x1} and \code{x2}.
}
}
\description{
This function calculates niche overlap between two species.
}
\details{
This function replicates the procedure presented in Broennimann, O., Fitzpatrick, M.C., Pearman, P.B., Petitpierre, B., Pellissier, L., Yoccoz, N.G., Thuiller, W., Fortin, M-J., Randin, C., Zimmermann, N.E., Graham, C.H., and Guisan, A.  2012.  Measuring ecological niche overlap from occurrence and spatial environmental data.  Global Ecology and Biogeography 21:481-497.
}
\examples{
data(lemur)
}
