% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictEnmSdm.r
\name{predictEnmSdm}
\alias{predictEnmSdm}
\title{Generic predict function for LMs, GLMs, GAMs, RFs, BRTs, CRFs, Maxent, and more}
\usage{
predictEnmSdm(model, newdata, maxentFun = "dismo", ...)
}
\arguments{
\item{model}{Object of class \code{lm}, \code{glm}, \code{gam}, \code{randomForest}, \code{MaxEnt}, \code{MaxNet}, and possibly others (worth a try!).}

\item{newdata}{Data frame or matrix with data to which to predict}

\item{maxentFun}{Either \code{'dismo'} in which case a Maxent model is predicted using the default \code{predict} function from the \pkg{dismo} package, or \code{'enmSdm'} in which case the function \code{\link[enmSdm]{predictMaxEnt}} function from the \pkg{enmSdm} package is used.}

\item{...}{Arguments to pass to the algorithm-specific \code{predict} function.}
}
\value{
Numeric.
}
\description{
This is a generic predict function that automatically uses the model common arguments for predicting models of the following types: linear models, generalized linear models (GLMs), generalized additive models (GAMs), random forests, boosted regression trees (BRTs)/gradient boosting machines (GBMs), conditional random forests, Maxent, and more.
}
\examples{

}
\seealso{
\code{\link[stats]{predict}} from the stats package, \code{\link[dismo]{predict}} from the dismo package, \code{\link[raster]{predict}} fromm the raster package
}
