% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRast.r
\name{sampleRast}
\alias{sampleRast}
\title{Sample random points from a raster with/out replacement}
\usage{
sampleRast(x, n, adjArea = TRUE, replace = TRUE, prob = TRUE)
}
\arguments{
\item{x}{Raster* object.}

\item{n}{Positive integer. Number of points to draw.}

\item{adjArea}{Logical. If TRUE then adjust probabilities so sampling accounts for cell area.}

\item{replace}{Logical. If TRUE then sample with replacement.}

\item{prob}{Logical. If TRUE then sample cells with probabilities proportional to cell values. If `adjArea` is also TRUE then probabilities are drawn proportional to the product of cell area * the value of the cell.}
}
\value{
2-column matrix with longitude and latitude of random points.
}
\description{
This function returns coordinates randomly located on a raster where cells can be sampled more than once if desired (sampled with replacement) and where the probability of selection is proportinate to the cell value (plus maybe cell area--both if desired).
}
\seealso{
\code{\link[dismo]{randomPoints}}, \code{\link{sampleRastStrat}}
r <- raster::raster()
nc <- raster::ncell(r)
r[] <- 1:nc
rands1 <- sampleRast(r, 10000)
rands2 <- sampleRast(r, 10000, adjArea=FALSE)
rands3 <- sampleRast(r, 10000, prob=FALSE)
rands4 <- sampleRast(r, 10000, adjArea=FALSE, prob=FALSE)
par(mfrow=c(2, 2))
plot(r, main='adjArea = TRUE & prob = TRUE')
points(rands1, pch='.')
plot(r, main='adjArea = FALSE & prob = TRUE')
points(rands2, pch='.')
plot(r, main='adjArea = TRUE & prob = FALSE')
points(rands3, pch='.')
plot(r, main='adjArea = FALSE & prob = FALSE')
points(rands4, pch='.')
}
