% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainLars.r
\name{trainLars}
\alias{trainLars}
\title{Model construction for Least Angle Regression (LARS) model}
\usage{
trainLars(data, resp = 1, preds = 2:ncol(data), alphas = c(0.01,
  seq(0.1, 1, by = 0.1)), scale = TRUE, quadratic = TRUE,
  cubic = TRUE, interaction = TRUE, interQuad = TRUE,
  na.rm = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{data}{Data frame.}

\item{resp}{Character or integer. Name or column index of response variable. Default is to use the first column in \code{data}.}

\item{preds}{Character list or integer list. Names of columns or column indices of predictors. Default is to use the second and subsequent columns in \code{data}.}

\item{alphas}{Numeric or numeric vector in the range (0, 1]. Degree of ridge penalty to impose (values close to 0 ==> full ridge penalty, while a value of 1 imposes no rifhe penalty).}

\item{scale}{Logical. If TRUE then scale values in \code{data[ , preds]} are rescaled to have mean of 0 and standard deviation of 1.}

\item{quadratic}{Logical. If TRUE then include quadratic terms in model construction stage for non-factor predictors. Quadratic columns will be named \code{<predictor name>_pow2}.}

\item{cubic}{Logical. If TRUE then include cubic terms in model construction stage for non-factor predictors. Cubic columns will be named \code{<predictor name>_pow3}.}

\item{interaction}{Logical. If TRUE then include 2-way interaction terms (including interactions between factor predictors). Interaction columns will be named \code{<predictor 1 name>_by_<predictor 2 name>}.}

\item{interQuad}{Logical. If TRUE then include all possible interactions of the form \code{x * y^2} unless \code{y} is a factor (linear-by-quadratic features). Linear-by-quadratic columns will be named \code{<predictor 1 name>_by_<predictor 2 name>_pow2}.}

\item{na.rm}{Logical. If TRUE then remove all rows of \code{data} in which there is at least one \code{NA} among \code{resp} or \code{preds}. The default is FALSE, which will cause an error if any row has an \code{NA}.}

\item{verbose}{Logical. If \code{TRUE} then display progress.}

\item{...}{Arguments to pass to \code{grpreg} \code{grpregOverlap}, and \code{cv.grpregOverlap}, especially \code{family} and \code{penalty}. \emph{Do not} include the \code{'group'} argument or \code{alpha} arguments.}
}
\value{
Object of class \code{grpreg} and \code{grpregOverlap}.
}
\description{
This function calculates the least angle regression (LARS) using possibly overlapping grouped covariates. The model is fit using cross validation (the \code{cv.grpregOverlap} function). The cross-validation is calculated across values of the \code{alpha}, which controls the degree of ridge penalty (\code{alpha ~0} (bit not = 0) imposes the full ridge penalty and \code{alpha) = 1} imposes no ridge penalty). Higher-order terms are constructed (e.g., quadratic, 2-way interaction, etc.) and fitted in a manner that respects mrginality (i.e., all lower order terms will have non-zero coefficients if a high-order term is used).
}
\details{
If \code{scale} is \code{TRUE} then predictors with zero variance will be removed from the data before the model is trained.
}
\seealso{
\code{\link{makeLarsData}}, \code{\link{predictLars}}, \code{\link[grpreg]{grpreg}}, \code{\link[grpregOverlap]{grpregOverlap}}, \code{\link[grpregOverlap]{cv.grpregOverlap}}
}
