% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xToCoords.r
\name{xToCoords}
\alias{xToCoords}
\title{Extract geographic coordinates from an object}
\usage{
xToCoords(x, longLat = NULL, out = "default")
}
\arguments{
\item{x}{A \code{data.frame}, \code{matrix}, \code{SpatialPoints*}, \code{SpatVector}, or\code{sf} object. If a data frame or matrix then the coordinate reference system is assumed to be unprojected (WGS84).}

\item{longLat}{Two-element character vector \emph{or} two-element integer vector. If \code{x} is a data frame then this should be a character list specifying the names of the fields in \code{x} \emph{or} a two-element vector of integers that correspond to longitude and latitude (in that order). For example, \code{c('long', 'lat')} or \code{c(1, 2)}. If \code{x} is a matrix then this is a two-element vector indicating the column numbers in \code{x} that represent longitude and latitude. For example, \code{c(1, 2)}. If \code{x} is a \code{SpatialPoints}, \code{SpatVector}, or \code{sf} object then this is ignored.}

\item{out}{Type of object to return. Options are:
\itemize{
    \item \code{'default'}: Same type as input
        \item \code{'sp'}: \code{SpatialPoints} object
        \item \code{'vect'}: \code{SpatVaector} object
        \item \code{'sf'}: \code{sf} object
}}
}
\value{
Object of class\code{SpatialPoints}, class\code{SpatVector}, class\code{sf}, a 2-column \code{matrix}, or a 2-column \code{data.frame}.
}
\description{
This function extracts geographic coordinates from a data frame, matrix, or \code{SpatialPoints}* object
}
