% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xToCoords.r
\name{xToCoords}
\alias{xToCoords}
\title{Extract geographic coordinates from data frame, matrix, or \code{SpatialPoints}* object}
\usage{
xToCoords(x, longLat = NULL, sp = TRUE)
}
\arguments{
\item{x}{A data frame, matrix, \code{SpatialPoints}, or \code{SpatialPointsDataFrame} object. If a data frame or matrix then the coordinate reference system is assumed to be unprojected (WGS84).}

\item{longLat}{Two-element character list \emph{or} two-element integer list. If \code{x} is a data frame then this should be a character list specifiying the names of the fields in \code{x} \emph{or} a two-element list of integers that correspond to longitude and latitude (in that order). For example, \code{c('long', 'lat')} or \code{c(1, 2)}. If \code{x} is a matrix then this is a two-element list indicating the column numbers in \code{x} that represent longitude and latitude. For example, \code{c(1, 2)}. If \code{x} is a \code{SpatialPoints} object then this is ignored.}

\item{sp}{Logical. If \code{TRUE} then return object of class\code{SpatialPoints}. If \code{FALSE} then return object of class \code{matrix}.}
}
\value{
Object of class\code{SpatialPoints} or a 2-column \code{matrix}.
}
\description{
This function extracts geographic coordinates from a data frame, matrix, or \code{SpatialPoints}* object
}
\seealso{
\code{\link[sp]{SpatialPoints}}
}
